feat: refactor data for landing pages
